pyproject.toml
[tool.poetry]
name = "startup-gift"
version = "0.1.0"
description = ""
authors = ["shahzod <qoshqulovshahzod9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-ckeditor-5 = "^0.2.15"
drf-spectacular = "^0.28.0"
django-mptt = "^0.16.0"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
redis = "^5.2.1"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
